+++ src/utils/injectProto/genLink.js
@@ -1,8 +1,10 @@
+import kebabCase from 'lodash/kebabCase';
+
 /**
  *  详情页链接
  *  @param {object} entry 条目
  *  @param {string} pathname 子路径
  */
 export default function(entry, pathname = '') {
-  return `/${this.state.entryProp}/${entry._id}${pathname}`;
+  return `/${kebabCase(this.state.entryProp)}/${entry._id}${pathname}`;
 };